refactor: complete C-to-C++ conversion and restructure project into engine/
- Convert all 6 remaining .c -> .cpp (gl_backend, gl_rsurf, gl_sprite, s_load, touch, img_main)
- Remove old HLSDK (half-life), replace with latest hlsdk-portable from FWGS master
- Delete .c originals, update Android.mk references, remove -std=c99
- Move NanoGL from android/jni/src/NanoGL/nanogl/ -> engine/nanogl/
- Merge root mainui/ into engine/mainui/ (keeping Android-specific files)
- Move android/ project -> engine/android/
- Move jni/ NDK build files -> engine/jni/
- Move hlsdk-portable -> engine/hlsdk/
- Delete old root directories: android/, mainui/, jni/
- Update all build files for new paths (Application.mk, Android.mk, build.gradle)
- Update vgui include path (utils/vgui/include -> utils/fake_vgui/include)
- Update android.yml workflow for new directory structure

+++ engine/mainui/miniutl/strtools.cpp
@@ -347,7 +347,7 @@ void V_SplitString2( const char *pString, const char * const *pSeparators, int n
 			int separatorLen = (int)strlen( pSeparators[iFirstSeparator] );
 			if ( pFirstSeparator > pCurPos || ( pFirstSeparator == pCurPos && bIncludeEmptyStrings ) )
 			{
-				outStrings.AddToTail( AllocString( pCurPos, (int)(pFirstSeparator-pCurPos) ) );
+				outStrings.AddToTail( AllocString( pCurPos, pFirstSeparator-pCurPos ) );
 			}
 
 			pCurPos = pFirstSeparator + separatorLen;
@@ -431,5 +431,5 @@ int V_StrTrim( char *pStr )
 		*pLastWhiteBlock = 0;
 	}
 
-	return (int)(pDest - pStart);
+	return pDest - pStart;
 }
